reafactor: Break up the NetworkVariableSerialization file
