--- com.unity.netcode.gameobjects/Runtime/NetworkVariable/NetworkVariableSerialization.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 2c6ef5fdf2e94ec3b4ce8086d52700b3
-timeCreated: 1650985453
