+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: d960ae6c5b8241aa9e2906b709095ea1
+timeCreated: 1718215841
