+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization/FallbackSerializer.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 288dbe7d1ff74860ae3552c034485538
+timeCreated: 1718219109
