+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization/INetworkVariableSerializer.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: f78e258ef55f4ee89bc3f24d67b8d242
+timeCreated: 1718218205
