+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization/NetworkVariableEquality.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 24b8352a975044509931bf684ccfdb82
+timeCreated: 1718219366
