+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization/NetworkVariableSerialization.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7a943170e35746e8913dd494d79bb63d
+timeCreated: 1718215899
