+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization/TypedILPPInitializers.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 65bdb3e11a9a412ab5e936a9c96a3da0
+timeCreated: 1718216842
