+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization/TypedSerializerImplementations.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: bbfa170e9dd448bbbe381ce38d5c139d
+timeCreated: 1718216671
