+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Serialization/UserNetworkVariableSerialization.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: b295a6756640488b9824d2ec6e26ddae
+timeCreated: 1718218272
